Redid the UI to be similar to ltk-template. This is a sketch!
However, I added some more stuff too. On the right of the pane
is a "specification properties" which is used for the entire specification
and all animations.

Then under it is a combo box and "new animation" button. You will be able to
select the animation you want to work with or create a new one.

Then, upon selecting the animation, the Frames Definition changes to be
for that animation. In the Frames Definition you will be able to
isnert/dup/delete frames of animation, set runtimes and if it should repeat,
etc, etc.

Inserting will insert just after the currently shown frame (or frame 0 if
there are no frames).

Dup will duplicate the current frame.

Delete will delete the current frame.

So far no functionality, just plain a straight up mockup.

File->Quit does work though.